backend/requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
influxdb==5.0.0
pkg-resources==0.0.0
pytest-runner==3.0
python-bittrex==0.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
